New rake task menu structure
Promote top level tasks to the top most menu rather than nesting them
under the namespace of the same name as a pseudo-task "all".  Namespaces
now include the colon in the submenu name.  For example, the test task
used to appear in the "test" submenu as "all" but now appears in the main
rake tasks menu, with the "test" submenu being renamed to "test:".