Provide :Rspec navigation command
I've resisted this a long time, in part because I didn't want to slide
down the slippery slope of providing navigation commands for plugins,
and partially because I was hedging between :Rspec proper and commands
like :Rspecmodel and :Rspeccontroller.  Now that we have a cheap way to
check for the presence of RSpec, the first issue is moot.  So let's see
how :Rspec works out.

Let's also try disabling the relevant navigation commands if no
Test::Unit tests are found.